requirements.txt
click
python-dotenv
flake8
black